requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.5.0
    # via pydantic
bleach==6.0.0
    # via kaggle
build==0.10.0
    # via pip-tools
certifi==2023.5.7
    # via
    #   kaggle
    #   requests
charset-normalizer==3.2.0
    # via requests
click==8.1.5
    # via pip-tools
idna==3.4
    # via requests
kaggle==1.5.15
    # via -r requirements.in
packaging==23.1
    # via build
pip-tools==7.0.0
    # via -r requirements.in
py4j==0.10.9.7
    # via pyspark
pydantic==2.0.3
    # via -r requirements.in
pydantic-core==2.3.0
    # via pydantic
pyproject-hooks==1.0.0
    # via build
pyspark==3.4.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via kaggle
python-slugify==8.0.1
    # via kaggle
pyyaml==6.0
    # via -r requirements.in
requests==2.31.0
    # via kaggle
six==1.16.0
    # via
    #   bleach
    #   kaggle
    #   python-dateutil
text-unidecode==1.3
    # via python-slugify
tqdm==4.65.0
    # via kaggle
typing-extensions==4.7.1
    # via
    #   pydantic
    #   pydantic-core
urllib3==2.0.3
    # via
    #   kaggle
    #   requests
webencodings==0.5.1
    # via bleach
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
